3.0.133.9

+++ modules/letsencrypt/libs/letsencryptlib.sh
@@ -271,7 +271,31 @@ function get_acme_certificate { ## for container
 
 function regenerate_letsencrypt {
     
+    if [ "$(systemctl is-active pound.service)" != "active" ]
+    then
+        err "Pound is not running!"
+        systemctl status pound.service --no-pager
+        
+        exit 99
+    fi
+    
+    if [ "$(systemctl is-active acme-server.service)" != "active" ]
+    then
+        err "Acme server is not running!"
+        systemctl status acme-server.service --no-pager
+        
+        exit 98
+    fi
+    
+    mkdir -p "$SC_DATASTORE_DIR/cert"
+    mkdir -p /etc/letsencrypt/live
+    
     msg "Regenerate letsencrypt certificates"
+    letsencrypt_main
+    
+    
+    return
+    
     for _C in $(lxc_ls)
     do
         rm -rf $SRV/$_C/rootfs/var/log/srvctl/letsencrypt.log
